Modified4 CinepalApi commit

+++ routes/api.php
@@ -65,11 +65,11 @@ Route::group([
     Route::put('/contact/{id}' , 'Admin\ContactController@update')->where('id' , '[1-9]+');
     Route::delete('/contact/{id}' , 'Admin\ContactController@destroy')->where('id' , '[1-9]+');
 
-    Route::get('/roles' , 'Admin\RoleController@index')->name('admin.roles-get');;
+    Route::get('/roles' , 'Admin\RoleController@index')->name('admin.getRoles');
 
-    Route::get('/roles/{id}/edit' , 'Admin\RoleController@edit')->where('id' , '[1-9]+');
-    Route::put('/roles/{id}' , 'Admin\RoleController@update')->where('id' , '[1-9]+');
-    Route::delete('/roles/{id}' , 'Admin\RoleController@destroy')->where('id' , '[1-9]+');
+    Route::get('/roles/{id}/edit' , 'Admin\RoleController@edit')->where('id' , '[1-9]+')->name('admin.getRoleEdit');
+    Route::put('/roles/{id}' , 'Admin\RoleController@update')->where('id' , '[1-9]+')->name('admin.updateRole');
+    Route::delete('/roles/{id}' , 'Admin\RoleController@destroy')->where('id' , '[1-9]+')->name('admin.deleteRole');
 
     Route::get('/actors' , 'Admin\ActorsController@index');
     Route::get('/actors/{id}/edit' , 'Admin\ActorsController@edit')->where('id' , '[1-9]+');
@@ -79,7 +79,7 @@ Route::group([
 
     Route::get('/users' , 'Admin\UserController@index');
     Route::get('/users/{id}/edit' , 'Admin\UserController@edit')->where('id' , '[1-9]+');
-    Route::post('/users' , 'Admin\UserController@store')->name('admin.users-post');
+    Route::post('/users' , 'Admin\UserController@store')->name('admin.addUser');
     Route::put('/users/{id}' , 'Admin\UserController@update')->where('id' , '[1-9]+');
     Route::delete('/users/{id}' , 'Admin\UserController@destroy')->where('id' , '[1-9]+');
 
@@ -121,4 +121,4 @@ Route::get('/movies/categories/{cat}/subcategory/{id}' , 'MoviesController@getMo
 Route::get('/actorss' , 'ActorsController@getAllActors');
 Route::post('/contactt' , 'ContactController@sendContact');
 
-Route::post('/roles' , 'Admin\RoleController@store')->name('admin.roles-post');
+Route::post('/roles' , 'Admin\RoleController@store')->name('admin.addRole');
